- updated the Auction Overview Cards - updated the Epoch Status Cards
