+++ src/assets/scss/components/_components.scss
@@ -14,8 +14,8 @@
 @import '../../../components/MultilayerPercentage/multilayerPercentage.styles.scss';
 @import '../../../components/NftPreview/nftPreview.styles.scss';
 @import '../../../components/NodesTable/nodesTable.styles.scss';
-@import '../../../components/Nodes/AuctionListCards/auctionListCards.styles.scss';
 @import '../../../components/Nodes/AuctionListTable/auctionListTables.styles.scss';
+@import '../../../components/Nodes/NodesEpochStatusCards/nodesEpochStatusCards.styles.scss';
 @import '../../../components/Nodes/NodesStatusPreview/nodesStatusPreview.styles.scss';
 @import '../../../components/NotificationsBar/notificationsBar.styles.scss';
 @import '../../../components/Pager/pager.styles.scss';
