--- src/components/Nodes/AuctionListCards/components/NodesOverviewCards/index.ts
@@ -1,1 +1,0 @@
-export * from './NodesOverviewCards';
