--- src/components/Nodes/AuctionListCards/components/index.ts
@@ -1,1 +1,0 @@
-export * from './NodesOverviewCards';
