--- src/components/Nodes/AuctionListCards/index.ts
@@ -1,1 +1,0 @@
-export * from './AuctionListCards';
