+++ src/components/Nodes/NodesEpochStatusCards/index.ts
@@ -1,0 +1,1 @@
+export * from './NodesEpochStatusCards';
