+++ src/components/Nodes/NodesEpochStatusCards/nodesEpochStatusCards.styles.scss
@@ -1,4 +1,4 @@
-.auction-list-cards {
+.epoch-status-cards {
   .time-container {
     font-family: var(--font-sans-serif);
     letter-spacing: -0.05em;
